Move _Proj4Projection out of _epsg
Move _Proj4Projection out of _epsg so that it doesn't have an implicit requirement for pyepsg any more (and it probably didn't really belong there anyway)
