+++ lib/cartopy/crs.py
@@ -1421,6 +1421,73 @@ def _find_gt(a, x):
     return a[0]
 
 
+_GLOBE_PARAMS = {'datum': 'datum',
+                 'ellps': 'ellipse',
+                 'a': 'semimajor_axis',
+                 'b': 'semiminor_axis',
+                 'f': 'flattening',
+                 'rf': 'inverse_flattening',
+                 'towgs84': 'towgs84',
+                 'nadgrids': 'nadgrids'}
+
+
+class _Proj4Projection(Projection):
+    def __init__(self, proj4_str, x0=-180., x1=180., y0=-90., y1=90.):
+        self.proj4_str = proj4_str.strip() #just in case
+
+        terms = [term.strip('+').split('=') for term in self.proj4_str.split(' ')]
+        globe_terms = filter(lambda term: term[0] in _GLOBE_PARAMS, terms)
+        globe = Globe(**{_GLOBE_PARAMS[name]: value for name, value in
+                              globe_terms})
+        other_terms = []
+        for term in terms:
+            if term[0] not in _GLOBE_PARAMS:
+                if len(term) == 1:
+                    other_terms.append([term[0], None])
+                else:
+                    other_terms.append(term)
+        super(_Proj4Projection, self).__init__(other_terms, globe)
+
+        # Convert lat/lon bounds to projected bounds.
+        # GML defines gmd:EX_GeographicBoundingBox as:
+        #   Geographic area of the entire dataset referenced to WGS 84
+        # NB. We can't use a polygon transform at this stage because
+        # that relies on the existence of the map boundary... the very
+        # thing we're trying to work out! ;-)
+        geodetic = Geodetic()
+        lons = np.array([x0, x0, x1, x1])
+        lats = np.array([y0, y1, y1, y0])
+        points = self.transform_points(geodetic, lons, lats)
+        x = points[:, 0]
+        y = points[:, 1]
+        self.bounds = (x.min(), x.max(), y.min(), y.max())
+
+
+    def __repr__(self):
+        return '_Proj4Projection({})'.format(self.proj4_str)
+
+    @property
+    def boundary(self):
+        x0, x1, y0, y1 = self.bounds
+        return sgeom.LineString([(x0, y0), (x0, y1), (x1, y1), (x1, y0),
+                                 (x0, y0)])
+
+    @property
+    def x_limits(self):
+        x0, x1, y0, y1 = self.bounds
+        return (x0, x1)
+
+    @property
+    def y_limits(self):
+        x0, x1, y0, y1 = self.bounds
+        return (y0, y1)
+
+    @property
+    def threshold(self):
+        x0, x1, y0, y1 = self.bounds
+        return min(x1 - x0, y1 - y0) / 100.
+
+
 def epsg(code):
     """
     Return the projection which corresponds to the given EPSG code.
@@ -1450,5 +1517,4 @@ def proj4(proj4_str, xmin=-180., ymin=-90., xmax=180., ymax=90.):
         A proj4 string doesn't contain the area of validity, so you can either
         specify the bounds, or have them default to the entire globe.
     """
-    import cartopy._epsg
-    return cartopy._epsg._Proj4Projection(proj4_str, xmin, xmax, ymin, ymax)
+    return _Proj4Projection(proj4_str, xmin, xmax, ymin, ymax)
